requirements.txt
requests==2.22.0
redis==3.3.11
rq==1.1.0
python-magic==0.4.15
Django==2.2.6
psycopg2-binary==2.8.4
djangorestframework==3.10.3
django-filter==2.2.0
dj-pagination==2.4.0
django-widget-tweaks==1.4.5
drf-dynamic-fields==0.3.1
gunicorn==19.9.0
django-extensions==2.2.5
django-redis==4.10.0
pyecharts==1.5.1
bottleneck==1.2.1
drf-yasg==1.17.0
coverage==4.5.4
pycodestyle==2.5.0
